Release new version with simple-icons v1.17.1 (#26)
* Update status badge in README.md

* Update Simple Icons to v1.14.0 (#20)

* Update SimpleIcons dependency to v1.12.0

* Fix build error due to new `.get` API function [REVERTED]

This solution is a bit of a hack. I think the real solution needs to be 
found up-stream.

* Update SimpleIcons dependency to v1.13.0

* Update SimpleIcons dependency to v1.14.0

* Run local server with test script (#23)

Run local server with test script using anywhere. This automatically builds the test-page and opens it for you in a local server.

Having it open in a local server prevents issues as described in https://github.com/simple-icons/simple-icons-font/issues/19

* Update simple-icons to v1.17.1 (#24)

* Use 'slug' property in build script (#25)

* Replace usage of titleToFilename by icon.slug

* Remove utilities module as it is no longer used

* [CI] Custom script for release stage

* version bump