Use SVGO to fix SVGs from SimpleIcons before creating the font
By using an option in the SVGO configuration to keep/put spaces after 
flag in the SVG path the build errors we get in this project when 
creating the font dissappear. This configuration is already in the main 
repo[1], but since not all contributors use th SVGO cli there are still 
some icons that have the problem.

For now we will redo the optimization on our side so we can always 
create the font.

Closes #29

--
1. 
https://github.com/simple-icons/simple-icons/blob/develop/.svgo.yml#L4