Add Verify workflow job to verify build:testpage
Add an extra job to the Verify Workflow to verify that the testpage can
at least be build (can't verify if it works just yet (*)). I made it so
that this new job depends on the old job. This is because this job
requires that the font can be build properly. Without this job-
dependency, both jobs may fail for the same reason which could be
"confusing"/take additional time to verify. The downside of this
approach is that the overall time of this Workflow is more often than
not longer.

(*): we could consider adding some tool to take a screenshot of the test
page and upload that as an artefact.

--
1. https://docs.github.com/en/free-pro-team@latest/actions/reference/workflow-syntax-for-github-actions#jobsjob_idneeds