Update formatting using Prettier
Prettier-pug first changed

   class=`si-${icon.cssClass}`

to

  class='si-${icon.cssClass}'

breaking the page as string interpolation was deleted. Which I changed
to

  class='si-' + ${icon.cssClass}

in which it than changed the quotes, resulting in

  class="si-" + ${icon.cssClass}

Not particularly happy about having to remove the string interpolation
AND the combination of single and double quotes in one file for no
reasong, but I prefer this over having to tell Renovate to ignore
Prettier-pug.